+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/ConversationScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.items
@@ -32,13 +31,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import com.wire.android.R
-import com.wire.android.model.UserAvatarData
-import com.wire.android.ui.common.UserProfileAvatar
 import com.wire.android.ui.common.bottomsheet.MenuModalSheetLayout
-import com.wire.android.ui.common.colorsScheme
-import com.wire.android.ui.common.conversationColor
 import com.wire.android.ui.common.dialogs.OngoingActiveCallDialog
-import com.wire.android.ui.common.dimensions
 import com.wire.android.ui.common.error.CoreFailureErrorDialog
 import com.wire.android.ui.common.snackbar.SwipeDismissSnackbarHost
 import com.wire.android.ui.common.topappbar.CommonTopAppBar
@@ -53,14 +47,19 @@ import com.wire.android.ui.home.conversations.ConversationSnackbarMessages.Error
 import com.wire.android.ui.home.conversations.ConversationSnackbarMessages.ErrorSendingAsset
 import com.wire.android.ui.home.conversations.ConversationSnackbarMessages.ErrorSendingImage
 import com.wire.android.ui.home.conversations.ConversationSnackbarMessages.OnFileDownloaded
+import com.wire.android.ui.home.conversations.call.ConversationCallViewModel
+import com.wire.android.ui.home.conversations.call.ConversationCallViewState
 import com.wire.android.ui.home.conversations.delete.DeleteMessageDialog
 import com.wire.android.ui.home.conversations.edit.EditMessageMenuItems
+import com.wire.android.ui.home.conversations.info.ConversationInfoViewModel
+import com.wire.android.ui.home.conversations.info.ConversationInfoViewState
+import com.wire.android.ui.home.conversations.messages.ConversationMessagesViewModel
+import com.wire.android.ui.home.conversations.messages.ConversationMessagesViewState
 import com.wire.android.ui.home.conversations.mock.getMockedMessages
 import com.wire.android.ui.home.conversations.model.AttachmentBundle
 import com.wire.android.ui.home.conversations.model.MessageContent
 import com.wire.android.ui.home.conversations.model.MessageSource
 import com.wire.android.ui.home.conversations.model.UIMessage
-import com.wire.android.ui.home.conversationslist.common.GroupConversationAvatar
 import com.wire.android.ui.home.messagecomposer.KeyboardHeight
 import com.wire.android.ui.home.messagecomposer.MessageComposeInputState
 import com.wire.android.ui.home.messagecomposer.MessageComposer
@@ -78,82 +77,90 @@ import okio.Path.Companion.toPath
 
 @Composable
 fun ConversationScreen(
-    conversationViewModel: ConversationViewModel,
+    messageComposerViewModel: MessageComposerViewModel,
+    conversationCallViewModel: ConversationCallViewModel,
+    conversationInfoViewModel: ConversationInfoViewModel,
+    conversationMessagesViewModel: ConversationMessagesViewModel,
     commonTopAppBarViewModel: CommonTopAppBarViewModel
 ) {
     val coroutineScope = rememberCoroutineScope()
     val showDialog = remember { mutableStateOf(ConversationScreenDialogType.NONE) }
 
     val startCallAudioPermissionCheck = StartCallAudioBluetoothPermissionCheckFlow {
-        conversationViewModel.navigateToInitiatingCallScreen()
+        conversationCallViewModel.navigateToInitiatingCallScreen()
     }
-    val uiState = conversationViewModel.conversationViewState
+    val uiState = messageComposerViewModel.conversationViewState
 
-    LaunchedEffect(conversationViewModel.savedStateHandle) {
-        conversationViewModel.checkPendingActions()
+    LaunchedEffect(messageComposerViewModel.savedStateHandle) {
+        messageComposerViewModel.checkPendingActions()
     }
 
     when (showDialog.value) {
         ConversationScreenDialogType.ONGOING_ACTIVE_CALL -> {
             OngoingActiveCallDialog(onJoinAnyways = {
-                conversationViewModel.navigateToInitiatingCallScreen()
+                conversationCallViewModel.navigateToInitiatingCallScreen()
                 showDialog.value = ConversationScreenDialogType.NONE
             }, onDialogDismiss = {
                 showDialog.value = ConversationScreenDialogType.NONE
             })
         }
+
         ConversationScreenDialogType.NO_CONNECTIVITY -> {
             CoreFailureErrorDialog(coreFailure = NetworkFailure.NoNetworkConnection(null)) {
                 showDialog.value = ConversationScreenDialogType.NONE
             }
         }
+
         ConversationScreenDialogType.NONE -> {}
     }
 
     ConversationScreen(
         conversationViewState = uiState,
-        onMessageChanged = conversationViewModel::onMessageChanged,
-        onSendButtonClicked = conversationViewModel::sendMessage,
-        onSendAttachment = conversationViewModel::sendAttachmentMessage,
-        onDownloadAsset = conversationViewModel::downloadOrFetchAssetToInternalStorage,
-        onImageFullScreenMode = conversationViewModel::navigateToGallery,
-        onBackButtonClick = conversationViewModel::navigateBack,
-        onDeleteMessage = conversationViewModel::showDeleteMessageDialog,
-        onStartCall = { startCallIfPossible(conversationViewModel, showDialog, startCallAudioPermissionCheck, coroutineScope) },
-        onJoinCall = conversationViewModel::joinOngoingCall,
-        onSnackbarMessage = conversationViewModel::onSnackbarMessage,
-        onSnackbarMessageShown = conversationViewModel::clearSnackbarMessage,
-        onDropDownClick = conversationViewModel::navigateToDetails,
-        tempCachePath = conversationViewModel.provideTempCachePath(),
-        onOpenProfile = conversationViewModel::navigateToProfile,
-        onUpdateConversationReadDate = conversationViewModel::updateConversationReadDate,
-        isConversationMember = conversationViewModel.isConversationMemberState,
+        conversationCallViewState = conversationCallViewModel.conversationCallViewState,
+        conversationInfoViewState = conversationInfoViewModel.conversationInfoViewState,
+        conversationMessagesViewState = conversationMessagesViewModel.conversationViewState,
+        onMessageChanged = messageComposerViewModel::onMessageChanged,
+        onSendButtonClicked = messageComposerViewModel::sendMessage,
+        onSendAttachment = messageComposerViewModel::sendAttachmentMessage,
+        onDownloadAsset = conversationMessagesViewModel::downloadOrFetchAssetToInternalStorage,
+        onImageFullScreenMode = messageComposerViewModel::navigateToGallery,
+        onBackButtonClick = messageComposerViewModel::navigateBack,
+        onDeleteMessage = messageComposerViewModel::showDeleteMessageDialog,
+        onStartCall = { startCallIfPossible(conversationCallViewModel, showDialog, startCallAudioPermissionCheck, coroutineScope) },
+        onJoinCall = conversationCallViewModel::joinOngoingCall,
+        onSnackbarMessage = messageComposerViewModel::onSnackbarMessage,
+        onSnackbarMessageShown = messageComposerViewModel::clearSnackbarMessage,
+        onDropDownClick = conversationInfoViewModel::navigateToDetails,
+        tempCachePath = messageComposerViewModel.provideTempCachePath(),
+        onOpenProfile = conversationInfoViewModel::navigateToProfile,
+        onUpdateConversationReadDate = messageComposerViewModel::updateConversationReadDate,
+        isConversationMember = messageComposerViewModel.isConversationMemberState,
         commonTopAppBarViewModel = commonTopAppBarViewModel
     )
 
     DeleteMessageDialog(
-        state = conversationViewModel.deleteMessageDialogsState,
-        actions = conversationViewModel.deleteMessageHelper
+        state = messageComposerViewModel.deleteMessageDialogsState,
+        actions = messageComposerViewModel.deleteMessageHelper
     )
     DownloadedAssetDialog(
-        downloadedAssetDialogState = conversationViewModel.conversationViewState.downloadedAssetDialogState,
-        onSaveFileToExternalStorage = conversationViewModel::onSaveFile,
-        onOpenFileWithExternalApp = conversationViewModel::onOpenFileWithExternalApp,
-        hideOnAssetDownloadedDialog = conversationViewModel::hideOnAssetDownloadedDialog
+        downloadedAssetDialogState = conversationMessagesViewModel.conversationViewState.downloadedAssetDialogState,
+        onSaveFileToExternalStorage = conversationMessagesViewModel::onSaveFile,
+        onOpenFileWithExternalApp = conversationMessagesViewModel::onOpenFileWithExternalApp,
+        hideOnAssetDownloadedDialog = conversationMessagesViewModel::hideOnAssetDownloadedDialog
     )
 }
 
 private fun startCallIfPossible(
-    conversationViewModel: ConversationViewModel,
+    conversationCallViewModel: ConversationCallViewModel,
     showDialog: MutableState<ConversationScreenDialogType>,
     startCallAudioPermissionCheck: CallingAudioRequestFlow,
     coroutineScope: CoroutineScope
 ) {
     coroutineScope.launch {
-        if (!conversationViewModel.hasStableConnectivity()) {
+        if (!conversationCallViewModel.hasStableConnectivity()) {
             showDialog.value = ConversationScreenDialogType.NO_CONNECTIVITY
         } else {
-            conversationViewModel.establishedCallConversationId?.let {
+            conversationCallViewModel.establishedCallConversationId?.let {
                 showDialog.value = ConversationScreenDialogType.ONGOING_ACTIVE_CALL
             } ?: run {
                 startCallAudioPermissionCheck.launch()
@@ -176,6 +183,9 @@ private fun StartCallAudioBluetoothPermissionCheckFlow(
 @Composable
 private fun ConversationScreen(
     conversationViewState: ConversationViewState,
+    conversationCallViewState: ConversationCallViewState,
+    conversationInfoViewState: ConversationInfoViewState,
+    conversationMessagesViewState: ConversationMessagesViewState,
     onMessageChanged: (String) -> Unit,
     onSendButtonClicked: () -> Unit,
     onSendAttachment: (AttachmentBundle?) -> Unit,
@@ -193,121 +203,105 @@ private fun ConversationScreen(
     onUpdateConversationReadDate: (String) -> Unit,
     isConversationMember: Boolean,
     commonTopAppBarViewModel: CommonTopAppBarBaseViewModel
-) {
+) = with(conversationViewState) {
     val conversationScreenState = rememberConversationScreenState()
 
-    with(conversationViewState) {
-        val connectionStateOrNull = (conversationDetailsData as? ConversationDetailsData.OneOne)?.connectionState
+    val connectionStateOrNull = (conversationInfoViewState.conversationDetailsData as? ConversationDetailsData.OneOne)?.connectionState
 
-        MenuModalSheetLayout(
-            sheetState = conversationScreenState.modalBottomSheetState,
-            coroutineScope = conversationScreenState.coroutineScope,
-            menuItems = EditMessageMenuItems(
-                isMyMessage = conversationScreenState.isSelectedMessageMyMessage(),
-                onCopyMessage = conversationScreenState::copyMessage,
-                onDeleteMessage = {
-                    conversationScreenState.hideEditContextMenu()
-                    onDeleteMessage(
-                        conversationScreenState.selectedMessage?.messageHeader!!.messageId,
-                        conversationScreenState.isSelectedMessageMyMessage()
-                    )
-                }
-            ),
-            content = {
-                BoxWithConstraints {
-                    val currentScreenHeight: Dp = with(LocalDensity.current) { constraints.maxHeight.toDp() }
-                    val fullScreenHeight: Dp = remember { currentScreenHeight }
-
-                    // when ConversationScreen is composed for the first time we do not know the height
-                    // until users opens the keyboard
-                    var keyboardHeight: KeyboardHeight by remember {
-                        mutableStateOf(KeyboardHeight.NotKnown)
-                    }
-
-                    // if the currentScreenHeight is smaller than the initial fullScreenHeight
-                    // and we don't know the keyboard height yet
-                    // calculated at the first composition of the ConversationScreen, then we know the keyboard size
-                    if (keyboardHeight is KeyboardHeight.NotKnown && currentScreenHeight < fullScreenHeight) {
-                        val difference = fullScreenHeight - currentScreenHeight
-                        if (difference > KeyboardHeight.DEFAULT_KEYBOARD_TOP_SCREEN_OFFSET)
-                            keyboardHeight = KeyboardHeight.Known(difference)
-                    }
-
-                    Scaffold(
-                        topBar = {
-                            Column {
-                                CommonTopAppBar(commonTopAppBarViewModel = commonTopAppBarViewModel as CommonTopAppBarViewModel)
-                                ConversationScreenTopAppBar(
-                                    title = conversationName.asString(),
-                                    avatar = {
-                                        when (conversationAvatar) {
-                                            is ConversationAvatar.Group ->
-                                                GroupConversationAvatar(
-                                                    color = colorsScheme().conversationColor(id = conversationAvatar.conversationId)
-                                                )
-                                            is ConversationAvatar.OneOne -> UserProfileAvatar(
-                                                UserAvatarData(
-                                                    asset = conversationAvatar.avatarAsset,
-                                                    availabilityStatus = conversationAvatar.status,
-                                                    connectionState = connectionStateOrNull
-                                                )
-                                            )
-                                            ConversationAvatar.None -> Box(modifier = Modifier.size(dimensions().userAvatarDefaultSize))
-                                        }
-                                    },
-                                    onBackButtonClick = onBackButtonClick,
-                                    onDropDownClick = onDropDownClick,
-                                    isDropDownEnabled = conversationViewState.conversationDetailsData !is ConversationDetailsData.None,
-                                    onSearchButtonClick = { },
-                                    onPhoneButtonClick = onStartCall,
-                                    hasOngoingCall = hasOngoingCall,
-                                    onJoinCallButtonClick = onJoinCall,
-                                    isUserBlocked = connectionStateOrNull == ConnectionState.BLOCKED
-                                )
-                            }
-                        },
-                        snackbarHost = {
-                            SwipeDismissSnackbarHost(
-                                hostState = conversationScreenState.snackBarHostState,
-                                modifier = Modifier.fillMaxWidth()
-                            )
-                        },
-                        content = { internalPadding ->
-                            Box(modifier = Modifier.padding(internalPadding)) {
-                                ConversationScreenContent(
-                                    keyboardHeight = keyboardHeight,
-                                    messages = messages,
-                                    lastUnreadMessage = lastUnreadMessage,
-                                    onMessageChanged = onMessageChanged,
-                                    messageText = conversationViewState.messageText,
-                                    onSendButtonClicked = onSendButtonClicked,
-                                    onShowContextMenu = conversationScreenState::showEditContextMenu,
-                                    onSendAttachment = onSendAttachment,
-                                    onDownloadAsset = onDownloadAsset,
-                                    onImageFullScreenMode = onImageFullScreenMode,
-                                    conversationState = conversationViewState,
-                                    onMessageComposerError = onSnackbarMessage,
-                                    onSnackbarMessageShown = onSnackbarMessageShown,
-                                    conversationScreenState = conversationScreenState,
-                                    isFileSharingEnabled = isFileSharingEnabled,
-                                    tempCachePath = tempCachePath,
-                                    isUserBlocked = connectionStateOrNull == ConnectionState.BLOCKED,
-                                    isConversationMember = isConversationMember,
-                                    onOpenProfile = onOpenProfile,
-                                    onUpdateConversationReadDate = onUpdateConversationReadDate
-                                )
-                            }
-                        }
-                    )
-                }
+    MenuModalSheetLayout(
+        sheetState = conversationScreenState.modalBottomSheetState,
+        coroutineScope = conversationScreenState.coroutineScope,
+        menuItems = EditMessageMenuItems(
+            isMyMessage = conversationScreenState.isSelectedMessageMyMessage(),
+            onCopyMessage = conversationScreenState::copyMessage,
+            onDeleteMessage = {
+                conversationScreenState.hideEditContextMenu()
+                onDeleteMessage(
+                    conversationScreenState.selectedMessage?.messageHeader!!.messageId,
+                    conversationScreenState.isSelectedMessageMyMessage()
+                )
             }
-        )
-    }
+        ),
+        content = {
+            BoxWithConstraints {
+                val currentScreenHeight: Dp = with(LocalDensity.current) { constraints.maxHeight.toDp() }
+                val fullScreenHeight: Dp = remember { currentScreenHeight }
+
+                // when ConversationScreen is composed for the first time we do not know the height
+                // until users opens the keyboard
+                var keyboardHeight: KeyboardHeight by remember {
+                    mutableStateOf(KeyboardHeight.NotKnown)
+                }
+
+                // if the currentScreenHeight is smaller than the initial fullScreenHeight,
+                // and we don't know the keyboard height yet
+                // calculated at the first composition of the ConversationScreen, then we know the keyboard size
+                if (keyboardHeight is KeyboardHeight.NotKnown && currentScreenHeight < fullScreenHeight) {
+                    val difference = fullScreenHeight - currentScreenHeight
+                    if (difference > KeyboardHeight.DEFAULT_KEYBOARD_TOP_SCREEN_OFFSET)
+                        keyboardHeight = KeyboardHeight.Known(difference)
+                }
+
+                Scaffold(
+                    topBar = {
+                        Column {
+                            CommonTopAppBar(commonTopAppBarViewModel = commonTopAppBarViewModel as CommonTopAppBarViewModel)
+                            ConversationScreenTopAppBar(
+                                conversationInfoViewState = conversationInfoViewState,
+                                onBackButtonClick = onBackButtonClick,
+                                onDropDownClick = onDropDownClick,
+                                isDropDownEnabled = conversationInfoViewState.conversationDetailsData !is ConversationDetailsData.None,
+                                onSearchButtonClick = { },
+                                onPhoneButtonClick = onStartCall,
+                                hasOngoingCall = conversationCallViewState.hasOngoingCall,
+                                onJoinCallButtonClick = onJoinCall,
+                                isUserBlocked = connectionStateOrNull == ConnectionState.BLOCKED
+                            )
+                        }
+                    },
+                    snackbarHost = {
+                        SwipeDismissSnackbarHost(
+                            hostState = conversationScreenState.snackBarHostState,
+                            modifier = Modifier.fillMaxWidth()
+                        )
+                    },
+                    content = { internalPadding ->
+                        Box(modifier = Modifier.padding(internalPadding)) {
+                            ConversationScreenContent(
+                                keyboardHeight = keyboardHeight,
+                                snackbarMessage = conversationViewState.snackbarMessage ?: conversationMessagesViewState.snackbarMessage,
+                                messages = conversationMessagesViewState.messages,
+                                lastUnreadMessage = conversationMessagesViewState.lastUnreadMessage,
+                                onMessageChanged = onMessageChanged,
+                                messageText = conversationViewState.messageText,
+                                onSendButtonClicked = onSendButtonClicked,
+                                onShowContextMenu = conversationScreenState::showEditContextMenu,
+                                onSendAttachment = onSendAttachment,
+                                onDownloadAsset = onDownloadAsset,
+                                onImageFullScreenMode = onImageFullScreenMode,
+                                conversationState = conversationViewState,
+                                onMessageComposerError = onSnackbarMessage,
+                                onSnackbarMessageShown = onSnackbarMessageShown,
+                                conversationScreenState = conversationScreenState,
+                                isFileSharingEnabled = isFileSharingEnabled,
+                                tempCachePath = tempCachePath,
+                                isUserBlocked = connectionStateOrNull == ConnectionState.BLOCKED,
+                                isConversationMember = isConversationMember,
+                                onOpenProfile = onOpenProfile,
+                                onUpdateConversationReadDate = onUpdateConversationReadDate
+                            )
+                        }
+                    }
+                )
+            }
+        }
+    )
 }
 
 @Suppress("LongParameterList")
 @Composable
 private fun ConversationScreenContent(
+    snackbarMessage: ConversationSnackbarMessages?,
     keyboardHeight: KeyboardHeight,
     messages: List<UIMessage>,
     lastUnreadMessage: UIMessage?,
@@ -330,22 +324,8 @@ private fun ConversationScreenContent(
     onUpdateConversationReadDate: (String) -> Unit
 ) {
     val coroutineScope = rememberCoroutineScope()
-    val context = LocalContext.current
 
-    conversationState.onSnackbarMessage?.let { messageCode ->
-        val (message, actionLabel) = getSnackbarMessage(messageCode)
-        LaunchedEffect(conversationState.onSnackbarMessage) {
-            val snackbarResult = conversationScreenState.snackBarHostState.showSnackbar(message = message, actionLabel = actionLabel)
-            when {
-                // Show downloads folder when clicking on Snackbar cta button
-                messageCode is OnFileDownloaded && snackbarResult == SnackbarResult.ActionPerformed -> {
-                    context.startActivity(Intent(DownloadManager.ACTION_VIEW_DOWNLOADS))
-                    onSnackbarMessageShown()
-                }
-                snackbarResult == SnackbarResult.Dismissed -> onSnackbarMessageShown()
-            }
-        }
-    }
+    SnackBarMessage(snackbarMessage, conversationState, conversationScreenState, onSnackbarMessageShown)
 
     val lazyListState = rememberSaveable(lastUnreadMessage, saver = LazyListState.Saver) {
         LazyListState(
@@ -390,6 +370,29 @@ private fun ConversationScreenContent(
         isConversationMember = isConversationMember,
         securityClassificationType = conversationState.securityClassificationType
     )
+}
+
+@Composable
+private fun SnackBarMessage(
+    snackbarMessage: ConversationSnackbarMessages?,
+    conversationState: ConversationViewState,
+    conversationScreenState: ConversationScreenState,
+    onSnackbarMessageShown: () -> Unit
+): Unit? = snackbarMessage?.let { messageCode ->
+    val (message, actionLabel) = getSnackbarMessage(messageCode)
+    val context = LocalContext.current
+    LaunchedEffect(conversationState.snackbarMessage) {
+        val snackbarResult = conversationScreenState.snackBarHostState.showSnackbar(message = message, actionLabel = actionLabel)
+        when {
+            // Show downloads folder when clicking on Snackbar cta button
+            messageCode is OnFileDownloaded && snackbarResult == SnackbarResult.ActionPerformed -> {
+                context.startActivity(Intent(DownloadManager.ACTION_VIEW_DOWNLOADS))
+                onSnackbarMessageShown()
+            }
+
+            snackbarResult == SnackbarResult.Dismissed -> onSnackbarMessageShown()
+        }
+    }
 }
 
 @Composable
@@ -467,10 +470,10 @@ fun MessageList(
 @Composable
 fun ConversationScreenPreview() {
     ConversationScreen(
-        conversationViewState = ConversationViewState(
-            conversationName = UIText.DynamicString("Some test conversation"),
-            messages = getMockedMessages(),
-        ),
+        conversationViewState = ConversationViewState(),
+        conversationCallViewState = ConversationCallViewState(),
+        conversationInfoViewState = ConversationInfoViewState(conversationName = UIText.DynamicString("Some test conversation")),
+        conversationMessagesViewState = ConversationMessagesViewState(messages = getMockedMessages()),
         onMessageChanged = {},
         onSendButtonClicked = {},
         onSendAttachment = {},
