+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/SnackbarMessages.kt
@@ -1,5 +1,7 @@
 package com.wire.android.ui.home.conversations
 
+// TODO(refactor): Split into smaller viewmodel-specific messages
+//                 For example, some for sending assets, some for downloadnig assets, etc.
 sealed class ConversationSnackbarMessages {
     object ErrorPickingAttachment : ConversationSnackbarMessages()
     object ErrorMaxImageSize : ConversationSnackbarMessages()
@@ -8,8 +10,8 @@ sealed class ConversationSnackbarMessages {
     object ErrorDownloadingAsset : ConversationSnackbarMessages()
     object ErrorOpeningAssetFile : ConversationSnackbarMessages()
     object ErrorDeletingMessage: ConversationSnackbarMessages()
-    class ErrorMaxAssetSize(val maxLimitInMB: Int) : ConversationSnackbarMessages()
-    class OnFileDownloaded(val assetName: String?) : ConversationSnackbarMessages()
+    data class ErrorMaxAssetSize(val maxLimitInMB: Int) : ConversationSnackbarMessages()
+    data class OnFileDownloaded(val assetName: String?) : ConversationSnackbarMessages()
 }
 
 sealed class MediaGallerySnackbarMessages {
